isolate event method in its own handler
